refactor(tests): moved tests
- Tests moved from test_class_properties.py to test_admin.py
